adding definitions of misc methods
